fix: resolve ~30 bugs in speedy-commons and speedy-core
speedy-commons:
- SpeedyNull.SPEEDY_NULL -> final (mutable singleton)
- SpeedyInt.asInt() NPE on null value
- 7 numeric validation rules: asLong() -> asInt() for ENUM_ORD
- FutureRule/PastRule: handle TIME + ZONED_DATE_TIME, injectable Clock
- EntityBuilder.field('')/keyField(''): assert -> IllegalArgumentException
- CommonUtilTest: assert -> assertEquals (test always passed without -ea)
- ExpansionPathTracker: defensive copy of input set
- JavaToSpeedy: throw ConversionException instead of silent SpeedyNull data loss
- ConditionFactory: 3+ segment paths throw clear BadRequestException
- CommonUtil.convertToEnum: null-ordinal guard, dead fallback removed
- MetaModelBuilder: message separator, IllegalStateException, dead loop removed
- DateRangeRule: null/malformed/inverted range guards
- JavaTypeRegistry: BigDecimal/BigInteger precision preserved
- SpeedyContext.types(): unmodifiable set
- EntityMetadataImpl: cached field sets, defensive copy of collections
- RegexRule/EmailRule: precompiled Pattern, stricter email regex
- Metadata classes: removed class-level @Setter (immutable post-build)
- ValidationUtils: IllegalArgumentException -> BadRequestException
- Speedy constants: public static final
- SpeedyQueryImpl: maxPageSize capped at 1000, reject <=0
- SpeedyResponseWriter.writeText(''): writes '' not null
- New: ForbiddenException (403), NotImplementedException (501)
- AGENTS.md: corrected 'zero dependencies' claim

speedy-core:
- CreateHandler: null saved entity -> continue (was NPEing)
- MetadataUtil.isKeyCompleteInEntity: has()-check + isNull()/isEmpty()
- DefaultQueryProcessor.wrap: preserve SpeedyHttpRuntimeException status
- StructureToQuery: multi-operator conditions throw (was silent data corruption)
- StructureToQuery: malformed \/\/\ throw instead of skip
- SpeedyUriContext.addToOrderList: propagate invalid \ exception
- SpeedyFactory: response.getWriter().flush() -> flushBuffer() with try/catch
- UriParserHandler: replaceAll -> replaceFirst anchored (was stripping from values)
- DefaultExceptionMapper: mask persistence 4xx messages, remove Hibernate string matching
- PermissionCheckHandler: 400 -> ForbiddenException (403)
- DeleteHandler: entity not found -> NotFoundException (404)
- DeleteHandler/CreateHandler: per-entity rethrow preserves actual status
- RecordToSpeedy: collection associations -> NotImplementedException (501)
- RecordToSpeedy: NoSuchElementException -> InternalServerError with entity name
- EventProcessor/ValidationProcessor: aligned isAssignableFrom direction
- DefaultQueryProcessor.create: empty check moved before insert
- CreateHandler/DeleteHandler: null e.getMessage() fallback
- SpeedyUriContext.parse: SpeedyHttpException passthrough
- MetaModelVerifier.verify: boolean -> void
- DefaultSpeedyEngine.queryProcessorCache: bounded LRU (was unbounded leak)
- RegistryImpl: LinkedList -> ArrayList
- RequestParserHandler: chunked transfer body always bounded
- AdviceExceptionMapper: failed handler -> ERROR + SpeedyHttpRuntimeException
- Deleted Hibernate stub test classes

Tests: 489 integration + all unit tests pass, 0 failures

+++ speedy-commons/src/main/java/com/github/silent/samurai/speedy/models/SpeedyInt.java
@@ -35,7 +35,7 @@ public class SpeedyInt implements SpeedyValue {
 
     @Override
     public Long asInt() {
-        return getValue().longValue();
+        return getValue();
     }
 
     @Override
