+++ speedy-commons/src/main/java/com/github/silent/samurai/speedy/validation/rules/DecimalMinRule.java
@@ -25,7 +25,7 @@ public class DecimalMinRule implements FieldRule {
         if (val == null || val.isEmpty()) return;
         if (!val.isNumber()) return;
 
-        BigDecimal num = val.isDouble() ? BigDecimal.valueOf(val.asDouble()) : BigDecimal.valueOf(val.asLong());
+        BigDecimal num = val.isDouble() ? BigDecimal.valueOf(val.asDouble()) : BigDecimal.valueOf(val.asInt());
         int cmp = num.compareTo(min);
         boolean valid = inclusive ? cmp >= 0 : cmp > 0;
         if (!valid) {
